+++ src/searchbar/searchbar.component.jsx
@@ -8,20 +8,15 @@ import {
 const ENTER = 13;
 
 class SearchBar extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = { value: this.props.value };
-  }
-
-  shouldComponentUpdate(nextProps, nextState) {
-    if (this.state.value !== nextState.value) {
+  shouldComponentUpdate(nextProps) {
+    if (this.props.value !== nextProps.value) {
       return true;
     }
     return false;
   }
 
   handleChange = (e) => {
-    this.setState({ value: e.target.value });
+    this.props.onChange(e.target.value);
   }
 
   handleOnKeyDown = (e) => {
@@ -31,29 +26,45 @@ class SearchBar extends React.Component {
   }
 
   handleSearch = () => {
-    if (this.props.action) {
-      this.props.action(this.state.value);
+    if (this.props.onSearch) {
+      this.props.onSearch(this.props.value);
     }
   }
 
   isValid = () => {
-    const isValid = (this.state.value || '').length > 0;
+    const isValid = (this.props.value || '').length > 0;
     return isValid;
   }
 
   render() {
     const isValid = this.isValid();
+    const style = {
+      height: 34,
+    };
+
     return (
       <InputGroup>
         <FormControl
           placeholder={this.props.placeholder}
           type="text"
-          value={this.state.value}
+          value={this.props.value}
           onChange={this.handleChange}
           onKeyDown={this.handleOnKeyDown}
         />
         <InputGroup.Button>
-          <Button onClick={this.handleSearch} disabled={!isValid}>Search</Button>
+          <Button onClick={this.handleSearch} disabled={!isValid} style={style}>
+            <svg
+              viewBox="0 0 16.83 16.46" width="17" height="17"
+            ><defs>
+              <style>
+                { `.oc-searchbar-icon
+                  {
+                    fill:#FFFFFF;
+                  }`
+                }
+              </style></defs><title>Search</title><path className="oc-searchbar-icon" d="M16.19,14.38,11.85,10A6.21,6.21,0,0,0,2.42,2,6.22,6.22,0,0,0,6.81,12.62a6.16,6.16,0,0,0,3.63-1.18l4.34,4.34a1,1,0,0,0,1.41-1.41Zm-12.36-5a4.21,4.21,0,1,1,3,1.24A4.19,4.19,0,0,1,3.83,9.39Z" />
+            </svg>
+          </Button>
         </InputGroup.Button>
       </InputGroup>
     );
@@ -62,12 +73,12 @@ class SearchBar extends React.Component {
 
 SearchBar.defaultProps = {
   placeholder: null,
-  horizontal: false,
   value: '',
 };
 
 SearchBar.propTypes = {
-  action: PropTypes.func.isRequired,
+  onChange: PropTypes.func.isRequired,
+  onSearch: PropTypes.func.isRequired,
   placeholder: PropTypes.string,
   value: PropTypes.string,
 };