+++ test/searchbar/searchbar.component.spec.jsx
@@ -6,19 +6,17 @@ import { expect } from 'chai';
 import { mount } from 'enzyme';
 import sinon from 'sinon';
 import {
-  ControlLabel,
   FormControl,
   Button,
-  Col,
 } from 'react-bootstrap';
 
 import { SearchBar } from '../../src/index';
 
-
 describe('SearchBar component', function describe() {
   function renderElement(props) {
     const defaultProps = {
-      action: sinon.spy(),
+      onSearch: sinon.spy(),
+      onChange: sinon.spy(),
       label: 'My search',
       placeholder: 'Type search...',
     };
@@ -28,11 +26,6 @@ describe('SearchBar component', function describe() {
     return wrapper;
   }
 
-  // it('should have correct title', function it() {
-  //   const wrapper = renderElement();
-  //   expect(wrapper.find(ControlLabel).text()).to.eql('My search');
-  // });
-
   it('should set placeholder text', function it() {
     const wrapper = renderElement();
     expect(wrapper.find(FormControl).props().placeholder).to.eql(
@@ -41,12 +34,62 @@ describe('SearchBar component', function describe() {
 
   it('should execute search', function it() {
     const props = {
-      action: sinon.spy(),
+      onSearch: sinon.spy(),
       value: 'Laptop',
     };
     const wrapper = renderElement(props);
     const searchButton = wrapper.find(Button);
     searchButton.simulate('click');
-    expect(props.action.called).to.be.true;
+    expect(props.onSearch.called).to.be.true;
+  });
+
+  it('should call change handler', function it() {
+    const props = {
+      onChange: sinon.spy(),
+      value: 'Laptop',
+    };
+    const wrapper = renderElement(props);
+    const formControl = wrapper.find(FormControl);
+    formControl.simulate('change', { target: { value: 'mouse' } });
+    expect(props.onChange.called).to.be.true;
+  });
+
+  it('should set value', function it() {
+    const props = {
+      value: 'Laptop',
+    };
+    const wrapper = renderElement(props);
+    const formControl = wrapper.find(FormControl);
+    expect(formControl.props().value).to.eql(props.value);
+  });
+
+  function testEnter(result, value = '') {
+    const props = {
+      onSearch: sinon.spy(),
+      value,
+    };
+    const ENTER = 13;
+
+    const wrapper = renderElement(props);
+    const formControl = wrapper.find(FormControl);
+    formControl.simulate('keyDown', { keyCode: ENTER });
+    expect(props.onSearch.called).to.eql(result);
+  }
+
+  it('should execute search with ENTER', function it() {
+    testEnter(true, 'Laptop');
+  });
+
+  it('should not execute search with ENTER', function it() {
+    testEnter(false);
+  });
+
+  it('search button should be disabled', function it() {
+    const props = {
+      value: '',
+    };
+    const wrapper = renderElement(props);
+    const searchButton = wrapper.find(Button);
+    expect(searchButton.props().disabled).to.be.true;
   });
 });